refactor: rename interfaces and adjust coverage
