+++ .husky/pre-commit
@@ -1,2 +1,3 @@
 npx lint-staged
 npm run type:check
+npm run test:coverage
