+++ package.json
@@ -7,9 +7,9 @@
     "build": "swc src -d dist --copy-files --strip-leading-paths",
     "start": "node dist/index.js",
     "dev": "nodemon --watch src --ext ts,js,json,md --exec node -r @swc-node/register src/index.ts",
-    "test": "vitest run",
-    "test:watch": "vitest",
-    "test:coverage": "vitest run --coverage",
+    "test": "vitest run --config vitest.config.ts",
+    "test:watch": "vitest --config vitest.config.ts",
+    "test:coverage": "vitest run --config vitest.config.ts --coverage",
     "type:check": "tsc --noEmit",
     "lint": "eslint . --ext .ts",
     "lint:fix": "npm run lint -- --fix",
