+++ src/repositories/interfaces/ChatUserRepository.interface.ts
