+++ src/repositories/interfaces/SummaryRepository.interface.ts
