+++ src/repositories/sqlite/SQLiteMessageRepository.ts
@@ -1,9 +1,9 @@
 import { inject, injectable } from 'inversify';
 
-import type { ChatMessage } from '../../services/ai/AIService';
-import type { StoredMessage } from '../../services/messages/StoredMessage';
+import type { ChatMessage } from '../../services/ai/AIService.interface';
+import type { StoredMessage } from '../../services/messages/StoredMessage.interface';
 import { DB_PROVIDER_ID, type SQLiteDbProvider } from '../DbProvider';
-import { type MessageRepository } from '../interfaces/MessageRepository';
+import { type MessageRepository } from '../interfaces/MessageRepository.interface';
 
 @injectable()
 export class SQLiteMessageRepository implements MessageRepository {
