+++ src/services/prompts/FilePromptService.ts
@@ -4,7 +4,7 @@ import { inject, injectable } from 'inversify';
 import { createLazy } from '../../utils/lazy';
 import { ENV_SERVICE_ID, EnvService } from '../env/EnvService';
 import { logger } from '../logging/logger';
-import { PromptService } from './PromptService';
+import { PromptService } from './PromptService.interface';
 
 @injectable()
 export class FilePromptService implements PromptService {
