+++ src/services/prompts/PromptService.interface.ts
