+++ src/services/summaries/SummaryService.interface.ts
