+++ src/triggers/InterestTrigger.ts
@@ -3,7 +3,7 @@ import { Context } from 'telegraf';
 import { DialogueManager } from '../services/chat/DialogueManager';
 import { InterestChecker } from '../services/interest/InterestChecker';
 import { logger } from '../services/logging/logger';
-import { Trigger, TriggerContext, TriggerResult } from './Trigger';
+import { Trigger, TriggerContext, TriggerResult } from './Trigger.interface';
 
 export class InterestTrigger implements Trigger {
   constructor(private checker: InterestChecker) {}
