+++ src/triggers/MentionTrigger.ts
@@ -2,7 +2,7 @@ import { Context } from 'telegraf';
 
 import { DialogueManager } from '../services/chat/DialogueManager';
 import { logger } from '../services/logging/logger';
-import { Trigger, TriggerContext, TriggerResult } from './Trigger';
+import { Trigger, TriggerContext, TriggerResult } from './Trigger.interface';
 
 export class MentionTrigger implements Trigger {
   async apply(
