+++ src/triggers/Trigger.interface.ts
