+++ vitest.config.ts
@@ -1,0 +1,20 @@
+/* eslint-disable import/no-unused-modules */
+import { defineConfig } from 'vitest/config';
+
+export default defineConfig({
+  test: {
+    coverage: {
+      provider: 'v8',
+      thresholds: { lines: 60, functions: 65, branches: 80 },
+      include: ['src/**/*.ts'],
+      exclude: [
+        'dist/**',
+        '**/index.ts',
+        'src/migrate.ts',
+        'src/repositories/**',
+        'src/services/env/**',
+        '**/*.interface.ts',
+      ],
+    },
+  },
+});
